create formatUtils with getNewDiscourseNodeText
